[zpr-ext] break up std into multiple files
