+++ zpr-ext/src/std/os.rs
@@ -1,0 +1,3 @@
+pub mod fd;
+#[cfg(any(doc, unix))]
+pub mod unix;
